# reset solver when copying parameters