--- a/src/dataobject.cpp
+++ b/src/dataobject.cpp
@@ -54,7 +54,7 @@ dataobject::dataobject(Rcpp::NumericMatrix _data,
   
   col.resize(8,0);
   
-  any_copy = parnames.size() > 0;
+  any_copy = par_from.size() > 0;
   done_copying = false;
   last_copy_row = -1;
   next_copy_row = 0;
@@ -88,7 +88,7 @@ dataobject::dataobject(Rcpp::NumericMatrix _data,
   
   col.resize(8,0);
   
-  any_copy = parnames.size() > 0;
+  any_copy = par_from.size() > 0;
   done_copying = false;
   last_copy_row = -1;
   next_copy_row = 0;
@@ -186,10 +186,10 @@ void dataobject::idata_row() {
 
 void dataobject::copy_parameters(int this_row, odeproblem* prob) {
   size_t n = par_from.size();
-  for(size_t i=0; i < n; ++i) {
+  for(size_t i = 0; i < n; ++i) {
     prob->param(par_to[i],Data(this_row,par_from[i]));
   }
-  prob->lsoda_init();
+  if(n > 0) prob->lsoda_init();
 }
 
 void dataobject::next_id(int id_n) {