Issue #745

- reset solver any time parameters are copied from the data
- see the issue; there were solver overshoot issues when time varying covariates changed parameter values; this is fixed by re-initializing solver whenever we copy parameters