@kylebaron 2020-11-11
@yizhang-yiz Yeah, I was trying to avoid doing this if possible b/c it does slow things down.  There has always been some inaccuracy around this, but honestly it seems to be worse since I switched to that C++ implementation of dlsoda.  It's hard to say but seems that way.  I'm not sure if there is anything we could look into in the solver or some way we could do what we need for the reset without giving up time step adaptation.  Is this worth looking into?