@yizhang-yiz 2020-11-11
>  Is this worth looking into?

In general yes but that'll depend on how much the two solvers diverges and the degree of slowdown. Since I can't think of any way to detect parameters that would cause unnecessary reinitialization, the current design makes sense to me.