@kylebaron 2020-11-13
@yizhang-yiz what do you think of this?  I've been benchmarking this a bit and this seems to take the edge off some of the slow down when parameters aren't changing from record to record (which I think is the main scenario where we could change things; if covariates are time-varying).