@yizhang-yiz 2020-11-13
It makes sense. What does the benchmark say about the performance difference?