@kylebaron 2020-11-13
# Before we changed anything


run a - `data1` is sparse input data, with constant covariate
run b - `data2` is sparse input data, without covariate
run c - `data3` is rich input with constant covariate
run d - `data4` is rich input data with no covariate

All runs have same duration and same number of output records.  There is a 
small penalty for rich input data (maybe more input processing).


```r
Unit: milliseconds
                    expr       min        lq      mean    median
 a <- mrgsim(mod, data1)  7.901581  8.800033 10.650774  9.310863
 b <- mrgsim(mod, data2)  7.968120  8.849370  9.741316  9.349273
 c <- mrgsim(mod, data3) 10.487445 12.019940 16.110642 12.752516
 d <- mrgsim(mod, data4) 10.191776 11.817085 14.856134 12.435446
```


# What we had yesterday

I made this to run slowly; simulated 3 days with every 15 minutes input record;
when there is a covariate with rich input(run c `data2`) we have to stop the 
solver every record and the time increases.  With sparse input, it doesn't make 
that big of a difference (it's like a handful of extra stops ... hard to detect).


```r
Unit: milliseconds
                    expr       min        lq      mean    median 
 a <- mrgsim(mod, data1)  7.741893  8.533279  9.791214  9.007484 
 b <- mrgsim(mod, data2)  7.795728  8.485864  9.774564  8.957434 
 c <- mrgsim(mod, data3) 35.744272 38.427414 44.985757 39.986618 
 d <- mrgsim(mod, data4) 10.219277 11.148762 13.466045 11.953646 
```



# With most recent update

With the most recent proposal; this works much faster because it doesn't restart
if the covariate value doesn't change.  If we had rich input with time-varying
covariate, you'd be back to the 44 ms run time.  But that seems to be the least
likely thing to happen.

But the penalty is that we have to do the extra work of comparing parameters 
comparison every time we update.


```r
Unit: milliseconds
                    expr       min        lq     mean    median 
 a <- mrgsim(mod, data1)  8.012418  8.991715 10.03263  9.538467 
 b <- mrgsim(mod, data2)  7.891747  8.980107 10.62298  9.505040 
 c <- mrgsim(mod, data3) 10.676537 12.227749 15.14205 12.990485 
 d <- mrgsim(mod, data4) 10.626696 11.882866 14.19894 12.798136 
```